third_party/blink/web_tests/fast/loader/javascript-url-hierarchical-execution-expected.txt

<rdar://problem/6904941> and https://bugs.webkit.org/show_bug.cgi?id=25300

This tests to see what WebKit does with javascript urls of the form "javascript://".
On the one hand, these are detected as hierarchical urls and usually marked invalid.
On the other hand, the contents after "javascript:" are valid javascript code - a comment - that should be executed.
If you click the link below and WebKit navigates - probably to an error page - then we're marking it as an invalid hierarchical URL and not trying to execute it.
Instead, clicking it should appear to do nothing and you should remain at this page.
In DumpRenderTree, if the custom policy delegate indicates that a load was attempted, then the test failed.

Click me to test
